:grapes: #90 photoUploadRepositoryを削除し、photoRepositoryに統合
